% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfpBayLasSte.R
\name{dfpBayLasSte}
\alias{dfpBayLasSte}
\title{Performs one iteration of DFP for the Bayesian Lasso}
\usage{
dfpBayLasSte(P, XX, Xy, yy, sN, hb, hs, hl, ht, hlsc, hlsh, nmcmc)
}
\arguments{
\item{P}{Current partition to be used.}

\item{XX}{Sufficient statistic X'X.}

\item{Xy}{Sufficient statistic X'y.}

\item{yy}{Sufficient statistic y'y.}

\item{sN}{Number of Observations.}

\item{hb}{Point estimate of the coefficients.}

\item{hs}{Point estimates of the variance of the error.}

\item{hl}{Point estimate for the global shrinking parameter.}

\item{ht}{Point estimate for the local shrinking parameters.}

\item{hlsc}{Hyper-prior for the lambda parameter rate.}

\item{hlsh}{Hyper-prior for the lambda parameter shape.}

\item{nmcmc}{Number of samples of the parameters.}
}
\value{
A list containing samples for the model parameters.
\describe{
  \item{sb}{A matrix with samples for coefficients, 1 sample per row.}
  \item{st}{A matrix with samples for local shrinkage parameters, 1 sample
   per row.}
  \item{ss}{Samples for sigma.}
  \item{sl}{Samples for the global shrinkage parameter.}
}
}
\description{
This function performs one step of the DFP with the Bayesian Lasso prior.
The inputs are described in Algorithm 1 of Guhaniyogi & Gutierrez.
}
\author{
Rene Gutierrez Marquez
}
