% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfpHorSho.R
\name{dfpHorSho}
\alias{dfpHorSho}
\title{Performs DFP for the Horseshoe Prior}
\usage{
dfpHorSho(
  staBat = 1,
  endBat = staBat,
  XDatPat = "./dat/datX-complete-",
  yDatPat = "./dat/daty-complete-",
  staPat = "./out/sta.RData",
  curParPat = "./out/curPar.RData",
  savPat = "./out/savPar.RData",
  suffix = "complete-dfp-BayLas",
  cores = parallel::detectCores()/2,
  staFla = TRUE,
  savCurPar = FALSE,
  nmcmc = 100,
  progress = TRUE,
  updPar = 1,
  staPar,
  savBol = TRUE,
  savCoe,
  datPer = seq(1:endBat),
  M,
  iniPar
)
}
\arguments{
\item{staBat}{Starting batch number.}

\item{endBat}{End batch number. As default is set as the starting batch.}

\item{XDatPat}{Path to the batch predictors. A default is provided as a
suggestion.}

\item{yDatPat}{Path to the batch responses.}

\item{staPat}{Path to the statistics of previous batches. If no file has
been created a new one is created with the path provided.}

\item{curParPat}{Path to the current parameter point estimates, partition and
sufficient statistics.}

\item{savPat}{Path to the file to save the parameters requested.}

\item{suffix}{Suffix to be added to the file paths.}

\item{cores}{Number of cores to be used in the parallel computations. Set as
default as half the number of cores detected.}

\item{staFla}{A boolean indicating if the statistics will be saved.}

\item{savCurPar}{A boolean indicating if the updated parameters will be
saved.}

\item{nmcmc}{Number of samples of the parameters.}

\item{progress}{A boolean indicating if a progress bar is desired. Default is
TRUE.}

\item{updPar}{Number of batches processed before the partition is updated.
Default is 1.}

\item{staPar}{Starting value for the parameters. Used only for the first
batch.}

\item{savBol}{Boolean indicating if the selected parameters will be saved.}

\item{savCoe}{Vector indicating the index of the coefficients to be saved.
This will also saved the associated parameters to each coefficient.}

\item{datPer}{Vector containing the order in which the data is to be
processed.}

\item{M}{Maximum number of elements in each part of the partition.}

\item{iniPar}{Initial partition.}
}
\value{
A list containing selected parameter samples, sufficient statistics,
and parameter partition.
\describe{
  \item{sb}{A matrix with samples for the selected coefficients, 1 sample per
  row.}
  \item{sl}{A matrix with samples for the selected local shrinkage
  parameters associated to the selected coefficients, 1 sample per row.}
  \item{sv}{A matrix with samples for the selected local shrinkage auxiliary
  hyper-parameters associated to the selected coefficients, 1 sample per row.}
  \item{ss}{Samples for sigma.}
  \item{st}{Samples for the global shrinkage parameter.}
  \item{se}{Samples for the auxilary global shrinkage hyper-parameter.}
  \item{XX}{Updated sufficient statistic X'X.}
  \item{Xy}{Updated sufficient statistic X'y.}
  \item{yy}{Updated sufficient statistic y'y.}
  \item{sN}{Updated sufficient statistic for the number of observations
  processed so far.}
  \item{P}{Updated partition.}
  \item{treshold}{Updated partition treshold correlation.}
}
}
\description{
This function runs DFP using the Horseshoe prior on data batches
previously stored in the indicated paths. It saves the performance statistics
indicated in the paper as well as the current point estimates and sufficient
statistics used for the following batch. It returns samples for selected
parameters for monitoring, as well as the updated sufficient statistics as
well as the updated partition and partition threshold parameter.
}
\author{
Rene Gutierrez Marquez
}
