% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfpHorShoSte.R
\name{dfpHorShoeSte}
\alias{dfpHorShoeSte}
\title{Performs one iteration of DFP for the Horseshoe Lasso}
\usage{
dfpHorShoeSte(P, XX, Xy, yy, sN, hb, hs, hl, ht, hv, he, nmcmc)
}
\arguments{
\item{P}{Current partition to be used.}

\item{XX}{Sufficient statistic X'X.}

\item{Xy}{Sufficient statistic X'y.}

\item{yy}{Sufficient statistic y'y.}

\item{sN}{Number of Observations.}

\item{hb}{Point estimate of the coefficients.}

\item{hs}{Point estimates of the variance of the error.}

\item{hl}{Point estimate for the local shrinking parameter.}

\item{ht}{Point estimate for the global shrinking parameters.}

\item{hv}{Point estimate for the auxilary local shrinking hyper-parameter.}

\item{he}{Point estimate for the auxilary global shrinking
hyper-parameters.}

\item{nmcmc}{Number of samples of the parameters.}
}
\value{
A list containing samples for the model parameters.
\describe{
  \item{sb}{A matrix with samples for coefficients, 1 sample per row.}
  \item{sl}{A matrix with samples for local shrinkage parameters, 1 sample
   per row.}
  \item{sv}{A matrix with samples for auxilary local shrinkage
  hyper-parameters, 1 sample per row.}
  \item{ss}{Samples for sigma.}
  \item{st}{Samples for the global shrinkage parameter.}
  \item{se}{Samples for the auxilary global shrinkage hyper-parameter.}
}
}
\description{
This function performs one step of the DFP with the Horseshoe Lasso prior.
The inputs are described in Algorithm 1 and the Supplementary Materials of
Guhaniyogi & Gutierrez.
}
\author{
Rene Gutierrez Marquez
}
