% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graPar.R
\name{graPar}
\alias{graPar}
\title{Partition creator for DFP}
\usage{
graPar(corMat, M)
}
\arguments{
\item{corMat}{Correlation Matrix.}

\item{M}{Maximum Number of Components.}
}
\value{
A list containing the Partition and the threshold determining it.
\describe{
  \item{P}{Partition.}
  \item{treshold}{Partition treshold correlation.}
}
}
\description{
Creates a partition from a graph based on a correlation matrix maximizing the
number of elements in each subset.
}
\author{
Rene Gutierrez Marquez
}
