% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{par}
\alias{par}
\title{Smallest Partition}
\usage{
par(P, M)
}
\arguments{
\item{P}{A list with a partition.}

\item{M}{Maximum Number of Components.}
}
\value{
A list with the new partition.
}
\description{
Creates a new partition with the smallest number of elements based on a
partition and a maximum number of elements per part.
}
\author{
Rene Gutierrez Marquez
}
