% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.sta.R
\name{pre.sta}
\alias{pre.sta}
\title{Statistics Creator}
\usage{
pre.sta(y, X, b, s)
}
\arguments{
\item{y}{Responses}

\item{X}{Predictors}

\item{b}{Samples for the coefficients.}

\item{s}{Samples for the variance of the error.}
}
\value{
A list containing selected parameter samples, sufficient statistics,
and parameter partition.
\describe{
  \item{preCreCov}{Prediction 95 Credible Interval Coverage.}
  \item{preSquErr}{Mean Prediction Square Error.}
  \item{preCreLen}{Prediction 95 Credible Interval Length.}
  \item{preIntSco}{Prediction 95 Credible Interval Score.}
}
}
\description{
This function creates the performance statistics used in Gutierrez &
Guhaniyogi.
}
\author{
Rene Gutierrez Marquez
}
